math: Fix cstrict error

+++ vlib/math/pow.v
@@ -116,7 +116,8 @@ pub fn pow(x f64, y f64) f64 {
 	// accumulate powers of two into exp.
 	mut x1, mut xe := frexp(x)
 	for i := i64(yi); i != 0; i >>= 1 {
-		if xe < -1 << 12 || 1 << 12 < xe {
+		// these series of casts is a little weird but we have to do them to prevent left shift of negative error
+		if xe < int(u32(u32(-1) << 12)) || 1 << 12 < xe {
 			// catch xe before it overflows the left shift below
 			// Since i !=0 it has at least one bit still set, so ae will accumulate xe
 			// on at least one more iteration, ae += xe is a lower bound on ae
